// src/store/oppfolging/action-type.ts
export enum OppfolgingActionType {
    HENT_OPPFOLGING = 'HENT_OPPFOLGING',
    HENT_OPPFOLGING_SUCCESS = 'HENT_OPPFOLGING_SUCCESS',
    HENT_OPPFOLGING_ERROR = 'HENT_OPPFOLGING_ERROR',
    START_ESKALERING = 'OPPRETTER_HENVENDELSE_SUCCESS',
    START_ESKALERING_SUCCESS = 'START_ESKALERING_SUCCESS',
    START_ESKALERING_ERROR = 'START_ESKALARING_ERROR',
    SETT_MANUELL = 'SETT_MANUELL',
    SETT_MANUELL_SUCCESS = 'SETT_MANUELL_SUCCESS',
    SETT_MANUELL_ERROR = 'SETT_MANUELL_ERROR',
    SETT_DIGITAL = 'SETT_DIGITAL',
    SETT_DIGITAL_SUCCESS = 'SETT_DIGITAL_SUCCESS',
    SETT_DIGITAL_ERROR = 'SETT_DIGITAL_ERROR',
    STOPP_KVP = 'STOPP_KVP',
    STOPP_KVP_SUCCESS = 'STOPP_KVP_SUCCESS',
    STOPP_KVP_ERROR = 'STOPP_KVP_ERROR',
    AVSLUTT_OPPFOLGING = 'AVSLUTT_OPPFOLGING',
    AVSLUTT_OPPFOLGING_SUCCESS = 'AVSLUTT_OPPFOLGING_SUCCESS',
    AVSLUTT_OPPFOLGING_ERROR = 'AVSLUTT_OPPFOLGING_ERROR',
    STOPP_ESKALERING = 'STOPP_ESKALERING',
    STOPP_ESKALERING_SUCCESS = 'STOPP_ESKALERING_SUCCESS',
    STOPP_ESKALERING_ERROR = 'STOPP_ESKALERING_ERROR',
}